screen.c: Leave Attribute Byte Untouched on Scroll
On scrolling, we copy the last row to penultimate row and set the last
row to be empty. We want to set zeros in character byte, but don't want
to overwrite background and foreground in attribute byte of video memory
cell.

This fixes the problem of invisible mouse cursor (in blank area) on the
scolled new lines.